requirements.txt
numpy
torch
hydra-core
omegaconf
scikit-learn
tensorboard
matplotlib
pillow
pyyaml
